+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - ex :e *path* opens file dialog at *path*
 - vcs execute and vcs_entry execute return process return code
 - list comparator is now a combobox
+- no more boost::process:v1 used
 
 ### Fixed